Update Yeoman.scala
able to define which gruntfile to use (in case you have a couple).

Note: This seems to compile but is untested. I think the string concat for building the gruntfile name is wrong. I only just started messing with sbt settings so i'm not sure what's the right way to do this. I'm sure it's an easy fix for anyone with some sbt settings knowledge.